+++ test/libsyncengine/jobs/network/testnetworkjobs.cpp
@@ -52,6 +52,7 @@
 #include "test_utility/localtemporarydirectory.h"
 #include "test_utility/remotetemporarydirectory.h"
 #include "test_utility/testhelpers.h"
+#include "test_utility/iohelpertests.h"
 #include "update_detection/file_system_observer/snapshot/snapshotitem.h"
 
 using namespace CppUnit;
@@ -142,7 +143,7 @@ void TestNetworkJobs::tearDown() {
     JobManager::stop();
     JobManager::clear();
     JobManager::reset();
-    MockIoHelperTestNetworkJobs::resetStdFunctions();
+    IoHelperTests::resetFunctions();
     TestBase::stop();
 }
 
@@ -409,8 +410,8 @@ void TestNetworkJobs::testDownload() {
                     ec = std::make_error_code(std::errc::cross_device_link);
 #endif
                 };
-        MockIoHelperTestNetworkJobs::setStdRename(MockRename);
-        MockIoHelperTestNetworkJobs::setStdTempDirectoryPath(MockTempDirectoryPath);
+        IoHelperTests::setRename(MockRename);
+        IoHelperTests::setTempDirectoryPathFunction(MockTempDirectoryPath);
 
         // CREATE
         {
@@ -449,7 +450,7 @@ void TestNetworkJobs::testDownload() {
             CPPUNIT_ASSERT(content == "test");
         }
 
-        MockIoHelperTestNetworkJobs::resetStdFunctions();
+        IoHelperTests::resetFunctions();
     }
 
     if (testhelpers::isRunningOnCI()) {
@@ -485,13 +486,10 @@ void TestNetworkJobs::testDownload() {
             const SyncPath localDestFilePath = temporaryDirectory.path() / "9Mo.txt";
             DownloadJob downloadJob(nullptr, _driveDbId, remoteTmpDir.id(), localDestFilePath, 0, 0, 0, false);
 
-            MockIoHelperTestNetworkJobs::setStdTempDirectoryPath([&smallPartitionPath](std::error_code &ec) {
-                ec.clear();
-                return smallPartitionPath;
-            });
+            IoHelperTests::setCacheDirectoryPath(smallPartitionPath);
 
             downloadJob.runSynchronously();
-            MockIoHelperTestNetworkJobs::resetStdFunctions();
+            IoHelperTests::resetFunctions();
             CPPUNIT_ASSERT_EQUAL_MESSAGE(std::string("Space available at " + smallPartitionPath.string() + " -> " +
                                                      std::to_string(Utility::getFreeDiskSpace(smallPartitionPath))),
                                          ExitInfo(ExitCode::SystemError, ExitCause::NotEnoughDiskSpace), downloadJob.exitInfo());
