+++ test/libsyncengine/jobs/network/testnetworkjobs.h
@@ -24,25 +24,6 @@
 using namespace CppUnit;
 
 namespace KDC {
-class MockIoHelperTestNetworkJobs : public IoHelper {
-    public:
-        static void setStdRename(std::function<void(const SyncPath &, const SyncPath &, std::error_code &)> rename) {
-            IoHelper::_rename = rename;
-        }
-        static void setStdTempDirectoryPath(std::function<SyncPath(std::error_code &)> tempDirectoryPath) {
-            IoHelper::_tempDirectoryPath = tempDirectoryPath;
-        }
-        static void resetStdFunctions() {
-            _isDirectory = static_cast<bool (*)(const SyncPath &path, std::error_code &ec)>(&std::filesystem::is_directory);
-            _isSymlink = static_cast<bool (*)(const SyncPath &path, std::error_code &ec)>(&std::filesystem::is_symlink);
-            _rename = static_cast<void (*)(const SyncPath &srcPath, const SyncPath &destPath, std::error_code &ecc)>(
-                    std::filesystem::rename);
-            _readSymlink = static_cast<SyncPath (*)(const SyncPath &path, std::error_code &ec)>(&std::filesystem::read_symlink);
-            _fileSize = static_cast<std::uintmax_t (*)(const SyncPath &path, std::error_code &ec)>(&std::filesystem::file_size);
-            _tempDirectoryPath = static_cast<SyncPath (*)(std::error_code &ec)>(&std::filesystem::temp_directory_path);
-        }
-};
-
 class TestNetworkJobs : public CppUnit::TestFixture, public TestBase {
     public:
         CPPUNIT_TEST_SUITE(TestNetworkJobs);
