+++ test/test_utility/iohelpertests.cpp
@@ -1,0 +1,73 @@
+/*
+ * Infomaniak kDrive - Desktop
+ * Copyright (C) 2023-2025 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "iohelpertests.h"
+
+namespace KDC {
+
+void IoHelperTests::setRename(std::function<void(const SyncPath &, const SyncPath &, std::error_code &)> f) {
+    _rename = f;
+}
+void IoHelperTests::setIsDirectoryFunction(std::function<bool(const SyncPath &path, std::error_code &ec)> f) {
+    _isDirectory = f;
+};
+void IoHelperTests::setIsSymlinkFunction(std::function<bool(const SyncPath &path, std::error_code &ec)> f) {
+    _isSymlink = f;
+};
+void IoHelperTests::setReadSymlinkFunction(std::function<SyncPath(const SyncPath &path, std::error_code &ec)> f) {
+    _readSymlink = f;
+};
+void IoHelperTests::setFileSizeFunction(std::function<std::uintmax_t(const SyncPath &path, std::error_code &ec)> f) {
+    _fileSize = f;
+}
+
+void IoHelperTests::setTempDirectoryPathFunction(std::function<SyncPath(std::error_code &ec)> f) {
+    _tempDirectoryPath = f;
+}
+
+void IoHelperTests::setCacheDirectoryPath(const SyncPath &newPath) {
+    IoError ioError = IoError::Success;
+    if (_previousCacheDirectoryPath.empty())
+        (void) IoHelper::cacheDirectoryPath(_previousCacheDirectoryPath, ioError); // Save the previous cache directory path
+    IoHelper::setCacheDirectoryPath(newPath);
+}
+
+#ifdef __APPLE__
+void IoHelperTests::setReadAliasFunction(std::function<bool(const SyncPath &path, SyncPath &targetPath, IoError &ioError)> f) {
+    _readAlias = f;
+};
+#endif
+
+void IoHelperTests::resetFunctions() {
+    // Reset to default std::filesytem implementation.
+    setIsDirectoryFunction(static_cast<bool (*)(const SyncPath &path, std::error_code &ec)>(&std::filesystem::is_directory));
+    setIsSymlinkFunction(static_cast<bool (*)(const SyncPath &path, std::error_code &ec)>(&std::filesystem::is_symlink));
+    setReadSymlinkFunction(static_cast<SyncPath (*)(const SyncPath &path, std::error_code &ec)>(&std::filesystem::read_symlink));
+    setFileSizeFunction(static_cast<std::uintmax_t (*)(const SyncPath &path, std::error_code &ec)>(&std::filesystem::file_size));
+    setTempDirectoryPathFunction(static_cast<SyncPath (*)(std::error_code &ec)>(&std::filesystem::temp_directory_path));
+
+#ifdef __APPLE__
+    // Default Utility::readAlias implementation
+    setReadAliasFunction([](const SyncPath &path, SyncPath &targetPath, IoError &ioError) -> bool {
+        std::string data;
+        return readAlias(path, data, targetPath, ioError);
+    });
+#endif
+    if (!_previousCacheDirectoryPath.empty()) IoHelper::setCacheDirectoryPath(_previousCacheDirectoryPath);
+}
+} // namespace KDC
